refactor(executor): move self-review and quality functions to separate file
- Move runSelfReview(), buildQualityGatesResult(), simpleQualityChecker struct and Check() method to self_review.go
- Self-review execution and quality gate result formatting now in dedicated module
- Improves separation of concerns, runSelfReview handles phase execution while quality.go has type definitions
- All tests passing, build verified

(GH-1138)
